+++ src/classes/redis-connection.ts
@@ -269,8 +269,8 @@ export class RedisConnection extends EventEmitter {
       if (lines[i].indexOf(maxMemoryPolicyPrefix) === 0) {
         const maxMemoryPolicy = lines[i].substr(maxMemoryPolicyPrefix.length);
         if (maxMemoryPolicy !== 'noeviction') {
-          throw new Error(
-            `Eviction policy is ${maxMemoryPolicy}. It should be "noeviction"`,
+          console.error(
+            `IMPORTANT! Eviction policy is ${maxMemoryPolicy}. It should be "noeviction"`,
           );
         }
       }