+++ src/types/backoff-strategy.ts
@@ -1,0 +1,8 @@
+import { Job } from '../classes/job';
+
+export type BackoffStrategy = (
+  attemptsMade?: number,
+  type?: string,
+  err?: Error,
+  job?: Job,
+) => Promise<number> | number;